easy

+++ entity/src/ip_address.rs
@@ -14,8 +14,8 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_metadata_id: Option<i32>,
+    pub id: i64,
+    pub ip_metadata_id: Option<i64>,
     pub addr: Option<String>,
     pub addr_type: Option<String>,
     pub created_at: Option<DateTimeWithTimeZone>,
@@ -38,7 +38,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -54,8 +54,8 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpMetadataId => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpMetadataId => ColumnType::BigInteger.def().null(),
             Self::Addr => ColumnType::Text.def().null(),
             Self::AddrType => ColumnType::Text.def().null(),
             Self::CreatedAt => ColumnType::TimestampWithTimeZone.def().null(),
