+++ entity/src/ip_hosting_details.rs
@@ -14,11 +14,11 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_main_id: Option<i32>,
-    pub num_domains: Option<i32>,
-    pub num_mail_servers: Option<i32>,
-    pub num_name_servers: Option<i32>,
+    pub id: i64,
+    pub ip_main_id: Option<i64>,
+    pub num_domains: Option<i64>,
+    pub num_mail_servers: Option<i64>,
+    pub num_name_servers: Option<i64>,
     pub created_at: Option<DateTimeWithTimeZone>,
     pub updated_at: Option<DateTimeWithTimeZone>,
 }
@@ -40,7 +40,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -55,11 +55,11 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpMainId => ColumnType::Integer.def().null(),
-            Self::NumDomains => ColumnType::Integer.def().null(),
-            Self::NumMailServers => ColumnType::Integer.def().null(),
-            Self::NumNameServers => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpMainId => ColumnType::BigInteger.def().null(),
+            Self::NumDomains => ColumnType::BigInteger.def().null(),
+            Self::NumMailServers => ColumnType::BigInteger.def().null(),
+            Self::NumNameServers => ColumnType::BigInteger.def().null(),
             Self::CreatedAt => ColumnType::TimestampWithTimeZone.def().null(),
             Self::UpdatedAt => ColumnType::TimestampWithTimeZone.def().null(),
         }
