+++ entity/src/ip_organization.rs
@@ -14,9 +14,9 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_main_id: Option<i32>,
-    pub asn: Option<i32>,
+    pub id: i64,
+    pub ip_main_id: Option<i64>,
+    pub asn: Option<i64>,
     pub domain: Option<String>,
     pub org_name: Option<String>,
     pub network: Option<String>,
@@ -46,7 +46,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -61,9 +61,9 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpMainId => ColumnType::Integer.def().null(),
-            Self::Asn => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpMainId => ColumnType::BigInteger.def().null(),
+            Self::Asn => ColumnType::BigInteger.def().null(),
             Self::Domain => ColumnType::Text.def().null(),
             Self::OrgName => ColumnType::Text.def().null(),
             Self::Network => ColumnType::Text.def().null(),
