+++ entity/src/ip_port.rs
@@ -14,8 +14,8 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_address_id: Option<i32>,
+    pub id: i64,
+    pub ip_address_id: Option<i64>,
     pub protocol: Option<String>,
     pub port_id: Option<String>,
     pub state: Option<String>,
@@ -40,7 +40,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -57,8 +57,8 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpAddressId => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpAddressId => ColumnType::BigInteger.def().null(),
             Self::Protocol => ColumnType::Text.def().null(),
             Self::PortId => ColumnType::Text.def().null(),
             Self::State => ColumnType::Text.def().null(),
