+++ entity/src/ip_security_flags.rs
@@ -14,8 +14,8 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_main_id: Option<i32>,
+    pub id: i64,
+    pub ip_main_id: Option<i64>,
     pub spam_database: Option<String>,
     pub adult_hosting: Option<bool>,
     pub hackers: Option<bool>,
@@ -42,7 +42,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -57,8 +57,8 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpMainId => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpMainId => ColumnType::BigInteger.def().null(),
             Self::SpamDatabase => ColumnType::Text.def().null(),
             Self::AdultHosting => ColumnType::Boolean.def().null(),
             Self::Hackers => ColumnType::Boolean.def().null(),
