+++ entity/src/ip_service.rs
@@ -14,8 +14,8 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_port_id: Option<i32>,
+    pub id: i64,
+    pub ip_port_id: Option<i64>,
     pub name: Option<String>,
     pub product: Option<String>,
     pub version: Option<String>,
@@ -48,7 +48,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -63,8 +63,8 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpPortId => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpPortId => ColumnType::BigInteger.def().null(),
             Self::Name => ColumnType::Text.def().null(),
             Self::Product => ColumnType::Text.def().null(),
             Self::Version => ColumnType::Text.def().null(),
