+++ entity/src/ip_stats.rs
@@ -14,8 +14,8 @@ impl EntityName for Entity {
 
 #[derive(Clone, Debug, PartialEq, DeriveModel, DeriveActiveModel, Eq, Serialize, Deserialize)]
 pub struct Model {
-    pub id: i32,
-    pub ip_metadata_id: Option<i32>,
+    pub id: i64,
+    pub ip_metadata_id: Option<i64>,
     pub up: Option<String>,
     pub down: Option<String>,
     pub total: Option<String>,
@@ -40,7 +40,7 @@ pub enum PrimaryKey {
 }
 
 impl PrimaryKeyTrait for PrimaryKey {
-    type ValueType = i32;
+    type ValueType = i64;
     fn auto_increment() -> bool {
         true
     }
@@ -55,8 +55,8 @@ impl ColumnTrait for Column {
     type EntityName = Entity;
     fn def(&self) -> ColumnDef {
         match self {
-            Self::Id => ColumnType::Integer.def(),
-            Self::IpMetadataId => ColumnType::Integer.def().null(),
+            Self::Id => ColumnType::BigInteger.def(),
+            Self::IpMetadataId => ColumnType::BigInteger.def().null(),
             Self::Up => ColumnType::Text.def().null(),
             Self::Down => ColumnType::Text.def().null(),
             Self::Total => ColumnType::Text.def().null(),
